add vrf03 benchamrks
