+++ Cargo.toml
@@ -24,8 +24,13 @@ thiserror = "1"
 
 [dev-dependencies]
 assert_cmd = "2.0"
+criterion = "0.3.5"
 hex = "0.4.2"
 predicates = "3.1"
 proptest = "1.6"
 rand = "0.8"
 tempfile = "3.19.1"
+
+[[bench]]
+name = "vrf03"
+harness = false
